pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[tool.flit.module]
name = "lightdash_client"

[project]
name = "lightdash-client-python"
authors = [{name = "yu-iskw"}]
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.8.0,<4"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Information Technology",
    "Intended Audience :: System Administrators",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Typing :: Typed",
]
dynamic = ["version", "description"]
dependencies = [
    "httpx >= 0.15.0, < 0.25.0",
    "attrs >= 21.3.0",
    "python-dateutil >= 2.8.0, < 3"
]

[project.urls]
Home = "https://github.com/yu-iskw/dbt-artifacts-parser"

[project.optional-dependencies]
test = [
    "pytest >=6.2.4,<7.0.0",
    "pylint >=2.12.0",
    "mypy ==0.910",
    "flake8 >=3.8.3,<4.0.0",
    "black ==21.9b0",
    "isort >=5.0.6,<6.0.0",
    "yapf >=0.29.0",
]
dev = [
    "flit ==3.7.1",
    "build ==0.7.0",
    "yapf >=0.29.0",
    "pyyaml >=5.3",
    "pdoc3 >=0.9.2",
    "pre-commit >=2.15.0",
]
